Use functional Astyle download URL
The previous download URL format, e.g.:

https://iweb.dl.sourceforge.net/project/astyle/astyle/astyle%203.0.1/astyle_3.1_linux.tar.gz

is currently giving a 443 response.

To find the new URL format, I employed the same technique as previously: download Artistic Style from the SourceForge
website and then check the URL it came from after being passed through the infamous download page. e.g.:

https://cfhcable.dl.sourceforge.net/project/astyle/astyle/astyle%203.0.1/astyle_3.1_linux.tar.gz

It might be subject to the same breakage over time, but if it works as long as the "iweb" URL did, then the approach is
still feasible.

I also considered using a Git clone of the repository, which would offer a stable URL at the expense of a larger
download. However, the project is using SVN and I don't have any interest in spending time on that.